cherry-pick async loading of Jitsi JS for react
thanks to @johnhalbert for the pointers!
